requirements.txt
numpy
pandas
matplotlib
tensorflow
keras
torch
opencv-python
pillow
jupyter
pdf2image
Levenshtein
os
easyocr
utils
preprocess
spellchecker
